Remove custom fnv partitioner and user the murmur2 from rdkafka
